requirements.txt
aiofile==3.8.8
aiofiles==23.2.1
alembic==1.11.3
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
caio==0.9.13
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.101.1
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
inflect==7.0.0
Mako==1.2.4
MarkupSafe==2.1.3
paramiko==3.3.1
passlib==1.7.4
psycopg2==2.9.7
pyasn1==0.5.0
pycparser==2.21
pydantic==2.2.1
pydantic-settings==2.0.3
pydantic_core==2.6.1
PyNaCl==1.5.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
sshtunnel==0.4.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
watchfiles==0.19.0
websockets==11.0.3
